% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webhook.R
\name{embeds}
\alias{embeds}
\alias{embed_image}
\title{Embed content with a message}
\usage{
embeds(description = NULL, ...)

embed_image(url, title = NULL)
}
\arguments{
\item{description}{Optional: Character string to use as description.}

\item{...}{Embedded content, i.e. \code{\link[=embed_image]{embed_image()}} objects.}

\item{url}{An image URL to embed.}

\item{title}{Title to be displayed above image.}
}
\value{

}
\description{
Multiple images and other content can be embedded in the same message,
which is why there's not just one big "embed image" that returns the
correct structure.
}
\examples{
\dontrun{
embeds = embeds(
  description = "Just another mtcars",
  embed_image(url = "https://i.imgur.com/TUdNWOW.png", title = "vroom")
)
}
}
